LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/url-encoded-expected.txt



PASS text.simple 
PASS textarea.simple 
FAIL nokeygen.simple assert_equals: expected "foo=barb" but got "foo=barb&=MIICQTCCASkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDARVnJmIe3Qtl7GF7JmoBE0RuEv9rnXkllqlAlYmk3O5OJO4lv6%2Forp7dv72KsRY4cXZ5j%2BE3rMgzfeYfBi%2B8%2FVT%2B5IXE7CkO9Zdwe%2BZ%2F3ys%2Ftmfjdu510HRPjnv4g7s1EOy5KGqpizej5qUggOqeGSxB5ccjXCIAebqPL7WfTH%2FIkCB%2BztEvUwcdhaN0bKUoEeB77Kk%2FSBbM22e16IxZzT0pFCK80kbaXB6GZeKTdhyvGCni4RCIvv56IGOWhkJ9HJeE4K0lf6yQkjSWe66OzW%2BJWGuJfi5XzaGxXSjeq8bXPVE%2BLGwumOEUEaGgmUWQMC%2F1yJI1w9VSlpMAVfEjNAgMBAAEWAQAwDQYJKoZIhvcNAQEEBQADggEBAD%2BmESJ2diq2OMAbW%2BnMekzOxuPdt92Ig%2FL9sp6wgY689vmeAh8cYWL0X2YxkMMPbDb4URDZ9oWeWdq6A%2BS%2FskrFkpg2K8NwU2rzthMPTS6Q4o1FWmCx9JX7XPMQ5sYO1W8RmLiIg84e69lmIPl53tN%2B%2BrEoDCTLjiOxv8PkvBuPqIDtseAYxjt%2FTrm%2FAG3FCfjksUGMZlFof1TVCYLiaXeQCyGDfJ3SIRSR%2FB%2FZHXkZmArPVPcFn%2F5AM36qn%2BoCZoQo%2FTv3zIRXqdYOAbgI%2FAADYczaT0GLcxlBeR0jyEKmR6irefSapP8OdnTKsF05fUnrvRRRgr63%2FBjN%2B99NefI%3D"
